// src/store/mutation-types.ts
export enum MutationTypes {
    SET_COLUMNS = 'setColumns',
    SET_ROWS = 'setRows',
    SET_BOARD = 'setBoard',
    SET_CURRENT_COLUMNS = 'setCurrentColumns',
    SET_PLAYER = 'setPlayer',
    SET_CURRENT_PLAYER = 'setCurrentPlayer',
    SET_IS_GAME_OVER = 'setIsGameOver',
    SET_WINNER = 'setWinner',
    RESET_STORE = 'resetStore',
}
  